# backend/docs/v2/v2_swagger.yaml
basePath: /api/v2
definitions:
...
  CancelExecutionRequest:
    properties:
      reason:
        maxLength: 500
        type: string
    required:
    - reason
    type: object
...
  CreateExecutionV2Request:
    properties:
      broker_reference:
        type: string
      ordered_amount:
        type: string
      ordered_quantity:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      error:
        type: string
    type: object
...
  ExecutionResponse:
    properties:
      broker_reference:
        type: string
      business_date:
        type: string
      cancellation_reason:
        type: string
      cancelled_at:
        type: string
      created_at:
        type: string
      currency:
        type: string
      decision_id:
        type: string
      executed_amount:
        type: string
      executed_at:
        type: string
      executed_quantity:
        type: string
      execution_price:
        type: string
      fund_id:
        type: string
      id:
        type: string
      instrument_code:
        type: string
      instrument_id:
        type: string
      ordered_amount:
        type: string
      ordered_quantity:
        type: string
      portfolio_id:
        type: string
      side:
        type: string
      status:
        type: string
      trader_user_id:
        type: string
      updated_at:
        type: string
    type: object
  FillExecutionRequest:
    properties:
      broker_reference:
        type: string
      executed_amount:
        type: string
      executed_quantity:
        type: string
      execution_price:
        type: string
      status:
        type: string
    type: object
...
  RecordConfirmationV2Request:
    properties:
      broker_reference:
        type: string
      confirmed_amount:
        type: string
      confirmed_price:
        type: string
      confirmed_quantity:
        type: string
      import_batch_id:
        type: string
    type: object
  ResolveConfirmationRequest:
    properties:
      discrepancy_reason:
        type: string
      target_status:
        enum:
        - MATCHED
        - MISMATCHED
        - REVIEWED
        type: string
    required:
    - target_status
    type: object
...
  TradeConfirmationResponse:
    properties:
      broker_reference:
        type: string
      business_date:
        type: string
      confirmed_amount:
        type: string
      confirmed_price:
        type: string
      confirmed_quantity:
        type: string
      created_at:
        type: string
      currency:
        type: string
      decision_id:
        type: string
      discrepancy_reason:
        type: string
      execution_id:
        type: string
      fund_id:
        type: string
      id:
        type: string
      import_batch_id:
        type: string
      portfolio_id:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Portfolio V2 endpoints (docs/api/portfolio-v2-api-ddd.md), mounted
    under /api/v2.
  title: IMS Thailand API — Portfolio V2
  version: 2.0.0
paths:
...
  /portfolios/{portfolioCode}/confirmations/{confirmationId}/resolve:
    post:
      consumes:
      - application/json
      description: Matches or reviews a trade confirmation that belongs to the resolved
        portfolio (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Confirmation UUID
        in: path
        name: confirmationId
        required: true
        type: string
      - description: Resolution fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResolveConfirmationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TradeConfirmationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve Portfolio Trade Confirmation By Code
      tags:
      - Investment - Portfolios V2
...
  /portfolios/{portfolioCode}/decisions/{decisionId}/executions:
    post:
      consumes:
      - application/json
      description: Opens an execution against an APPROVED decision that belongs to
        the resolved portfolio (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Decision UUID
        in: path
        name: decisionId
        required: true
        type: string
      - description: Execution fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateExecutionV2Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ExecutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: COMPLIANCE_NOT_CONFIGURED, COMPLIANCE_UNAVAILABLE, or evaluated
            rule rejection
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Portfolio Execution By Code
      tags:
      - Investment - Portfolios V2
...
  /portfolios/{portfolioCode}/executions/{executionId}/cancel:
    post:
      consumes:
      - application/json
      description: Cancels an execution that belongs to the resolved portfolio (Portfolio
        V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Execution UUID
        in: path
        name: executionId
        required: true
        type: string
      - description: Cancellation reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CancelExecutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ExecutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel Portfolio Execution By Code
      tags:
      - Investment - Portfolios V2
  /portfolios/{portfolioCode}/executions/{executionId}/confirmations:
    post:
      consumes:
      - application/json
      description: Records a broker confirmation against an execution that belongs
        to the resolved portfolio (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Execution UUID
        in: path
        name: executionId
        required: true
        type: string
      - description: Confirmation fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RecordConfirmationV2Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/TradeConfirmationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record Portfolio Trade Confirmation By Code
      tags:
      - Investment - Portfolios V2
  /portfolios/{portfolioCode}/executions/{executionId}/fill:
    post:
      consumes:
      - application/json
      description: Records a fill against an execution that belongs to the resolved
        portfolio (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Execution UUID
        in: path
        name: executionId
        required: true
        type: string
      - description: Fill fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/FillExecutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ExecutionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fill Portfolio Execution By Code
      tags:
      - Investment - Portfolios V2
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"